fix: restore sort icon state, add dragndropdesign mode and react icon
