+++ packages/pluggableWidgets/datagrid-web/src/components/icons/FaGripVertical.tsx
@@ -1,0 +1,15 @@
+import { ReactElement } from "react";
+
+/**
+ * Custom drag handle icon with 6 aligned dots in 3 rows
+ */
+export function FaGripVertical(): ReactElement {
+    return (
+        <svg aria-hidden="true" focusable="false" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 320 512">
+            <path
+                fill="currentColor"
+                d="M96 64c17.7 0 32 14.3 32 32s-14.3 32-32 32-32-14.3-32-32 14.3-32 32-32zm128 0c17.7 0 32 14.3 32 32s-14.3 32-32 32-32-14.3-32-32 14.3-32 32-32zm-128 160c17.7 0 32 14.3 32 32s-14.3 32-32 32-32-14.3-32-32 14.3-32 32-32zm128 0c17.7 0 32 14.3 32 32s-14.3 32-32 32-32-14.3-32-32 14.3-32 32-32zm-128 160c17.7 0 32 14.3 32 32s-14.3 32-32 32-32-14.3-32-32 14.3-32 32-32zm128 0c17.7 0 32 14.3 32 32s-14.3 32-32 32-32-14.3-32-32 14.3-32 32-32z"
+            ></path>
+        </svg>
+    );
+}
